backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
python-whois>=0.8.0
joblib>=1.3.0
scikit-learn==1.2.2
pydantic>=2.0.0
